+++ modules/MSP365/Public/Get-MSPCommand.ps1
@@ -1,4 +1,4 @@
-function Get-M365Commands {
+function Get-MSPCommand {
 
     Get-Command -Module MSP365*