+++ modules/MSP365.Authentication/ChangeLog.md
@@ -1,3 +1,4 @@
 # **MSP365.Reporting**
 
 - **1.0.0** Created the module
+- **1.0.1** Add new function Get-MSPAuthenticationContext