+++ modules/MSP365/ChangeLog.md
@@ -1,3 +1,4 @@
 # **MSP365**
 
 - **1.0.19** Add new module MSP365.Authentication
+- **1.0.20** Add new function Update-MSPModule and Get-MSPCommand